# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Remedy Scribe API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "description": "P0-8: multi-factor authentication for clinician access \u2014 password\nAND a valid TOTP code are both required; either failing alone returns\nthe same 401 to avoid leaking which factor was wrong.\n\nPhase 0.3: also rate-limited per-IP and locked-out per-email (see\napp/services/auth_rate_limit.py), and now issues a refresh token\nalongside the access token (see app/services/refresh_tokens.py) \u2014\nthe fix for \"a doctor gets logged out mid-consultation every 15\nminutes with no way back in short of re-entering a password and MFA\ncode.\"",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Refresh",
        "description": "Silent-renewal endpoint the mobile client calls when its access\ntoken is expired (or about to be) \u2014 no password/MFA re-entry, no\nAuthorization header. Rotates the refresh token on every use; a\nreplayed/stolen refresh token gets caught here, not at login.",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout",
        "description": "Revokes exactly the one refresh token presented \u2014 signing out\nthis device without touching any other session the clinician has.",
        "operationId": "logout_api_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/mfa/enroll": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Mfa Enroll",
        "description": "Step 1 of 2 (Phase 0.3) \u2014 replaces \"the TOTP secret can only be\ncreated by a seed script.\" Only usable on an account with no active\nMFA yet: an already-enrolled clinician re-provisioning here would\nlet anyone who has since learned the password silently swap out a\nreal MFA secret for their own. Re-enrollment for an already-active\naccount is intentionally not self-service (see\ndocs/decisions/0007) \u2014 it needs an admin/support path, not built\nhere.",
        "operationId": "mfa_enroll_api_v1_auth_mfa_enroll_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaEnrollRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MfaEnrollOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/mfa/enroll/confirm": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Mfa Enroll Confirm",
        "description": "Step 2 of 2: one valid code against the *pending* secret moves it\nto `mfa_secret`, the one login actually checks. Until this succeeds,\n/mfa/enroll can be called again (e.g. the QR was mis-scanned) \u2014 a\npending secret is disposable; only the confirmed one matters.",
        "operationId": "mfa_enroll_confirm_api_v1_auth_mfa_enroll_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaEnrollConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MfaEnrollConfirmOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/clinicians/{clinician_id}/revoke-sessions": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Revoke Sessions",
        "operationId": "revoke_sessions_api_v1_auth_clinicians__clinician_id__revoke_sessions_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "clinician_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Clinician Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokeSessionsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "mfa_code": {
            "type": "string",
            "title": "Mfa Code"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password",
          "mfa_code"
        ],
        "title": "LoginRequest"
      },
      "LogoutRequest": {
        "properties": {
          "refresh_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "title": "LogoutRequest",
        "description": "Single-session logout: revokes exactly the refresh token\npresented, leaving any other device's session untouched. Optional for\nthe same reason as RefreshRequest \u2014 the cookie is the browser path."
      },
      "MfaEnrollConfirmOut": {
        "properties": {
          "enrolled": {
            "type": "boolean",
            "title": "Enrolled"
          }
        },
        "type": "object",
        "required": [
          "enrolled"
        ],
        "title": "MfaEnrollConfirmOut"
      },
      "MfaEnrollConfirmRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "code": {
            "type": "string",
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password",
          "code"
        ],
        "title": "MfaEnrollConfirmRequest",
        "description": "Step 2: one valid code against the pending secret activates it."
      },
      "MfaEnrollOut": {
        "properties": {
          "provisioning_uri": {
            "type": "string",
            "title": "Provisioning Uri"
          }
        },
        "type": "object",
        "required": [
          "provisioning_uri"
        ],
        "title": "MfaEnrollOut"
      },
      "MfaEnrollRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "MfaEnrollRequest",
        "description": "Step 1 of enrollment: proves password ownership, then provisions\na *pending* secret \u2014 not yet the one login checks against."
      },
...
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "title": "RefreshRequest",
        "description": "Phase 0.3: exchanges a still-valid refresh token for a new\naccess/refresh pair. Deliberately takes no Authorization header \u2014\nthe access token this call is meant to replace may already be\nexpired, which is the whole reason to call it.\n\nPhase 2.1: optional, because the browser client (decision 0024) never\nsees the refresh token \u2014 it rides in an httpOnly cookie the route\nreads instead. The field stays for non-browser callers and for the\nPhase 0.3 tests that exercise rotation/reuse-detection directly."
      },
      "RevokeSessionsOut": {
        "properties": {
          "revoked_count": {
            "type": "integer",
            "title": "Revoked Count"
          }
        },
        "type": "object",
        "required": [
          "revoked_count"
        ],
        "title": "RevokeSessionsOut"
      },
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token"
        ],
        "title": "TokenResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}